LPS-105024 Move types to API and remove Generator interface for now
