+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-registry-api/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
 	compileOnly group: "org.json", name: "json", version: "20180813"
 	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
-	compileOnly project(":apps:portal-security-crypto:portal-security-crypto-generator-spi")
+	compileOnly project(":apps:portal-security-crypto:portal-security-crypto-generator-api")
 }
