+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-registry-impl/build.gradle
@@ -2,6 +2,7 @@ dependencies {
 	compileOnly group: "org.json", name: "json", version: "20180813"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:portal-security-crypto:portal-security-crypto-generator-api")
 	compileOnly project(":apps:portal-security-crypto:portal-security-crypto-generator-registry-api")
 	compileOnly project(":apps:portal-security-crypto:portal-security-crypto-generator-spi")
 	compileOnly project(":core:osgi-service-tracker-collections")
